+++ collect_app/src/androidTest/java/org/odk/collect/android/utilities/InstanceUploaderUtilsTest.java
@@ -11,6 +11,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.dao.InstancesDao;
+import org.odk.collect.android.dao.helpers.InstancesDaoHelper;
 import org.odk.collect.android.dto.Instance;
 import org.odk.collect.android.provider.InstanceProviderAPI;
 
@@ -60,7 +61,7 @@ public class InstanceUploaderUtilsTest {
                     .status(InstanceProviderAPI.STATUS_COMPLETE)
                     .lastStatusChangeDate(time)
                     .build();
-            instancesDao.saveInstance(DatabaseUtils.getValuesFromInstanceObject(instance));
+            instancesDao.saveInstance(InstancesDaoHelper.getValuesFromInstanceObject(instance));
         }
     }
