+++ collect_app/src/main/java/org/odk/collect/android/dao/helpers/ContentResolverHelper.java
@@ -25,7 +25,7 @@ import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.logic.FormInfo;
 import org.odk.collect.android.provider.FormsProviderAPI;
 import org.odk.collect.android.provider.InstanceProviderAPI;
-import org.odk.collect.android.utilities.DatabaseUtils;
+import org.odk.collect.android.utilities.FileUtils;
 
 public final class ContentResolverHelper {
 
@@ -68,7 +68,7 @@ public final class ContentResolverHelper {
         try (Cursor c = getContentResolver().query(uri, null, null, null, null)) {
             if (c != null && c.getCount() == 1) {
                 c.moveToFirst();
-                formPath = DatabaseUtils.getAbsoluteFilePath(c.getString(c.getColumnIndex(FormsProviderAPI.FormsColumns.FORM_FILE_PATH)));
+                formPath = FileUtils.getAbsoluteFilePath(c.getString(c.getColumnIndex(FormsProviderAPI.FormsColumns.FORM_FILE_PATH)));
             }
         }
         return formPath;
