+++ collect_app/src/main/java/org/odk/collect/android/dao/helpers/InstancesDaoHelper.java
@@ -14,15 +14,20 @@
 
 package org.odk.collect.android.dao.helpers;
 
+import android.content.ContentValues;
 import android.database.Cursor;
 import android.net.Uri;
 
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.dao.InstancesDao;
+import org.odk.collect.android.dto.Instance;
 import org.odk.collect.android.logic.FormController;
 import org.odk.collect.android.preferences.GeneralSharedPreferences;
 import org.odk.collect.android.preferences.GeneralKeys;
 import org.odk.collect.android.provider.InstanceProviderAPI;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import timber.log.Timber;
 
@@ -91,5 +96,70 @@ public final class InstancesDaoHelper {
             }
         }
         return isAvailable;
+    }
+
+    /**
+     * Returns all instances available through the cursor and closes the cursor.
+     */
+    public static List<Instance> getInstancesFromCursor(Cursor cursor) {
+        List<Instance> instances = new ArrayList<>();
+        if (cursor != null) {
+            try {
+                cursor.moveToPosition(-1);
+                while (cursor.moveToNext()) {
+                    int displayNameColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.DISPLAY_NAME);
+                    int submissionUriColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.SUBMISSION_URI);
+                    int canEditWhenCompleteIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.CAN_EDIT_WHEN_COMPLETE);
+                    int instanceFilePathIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.INSTANCE_FILE_PATH);
+                    int jrFormIdColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.JR_FORM_ID);
+                    int jrVersionColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.JR_VERSION);
+                    int statusColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.STATUS);
+                    int lastStatusChangeDateColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.LAST_STATUS_CHANGE_DATE);
+                    int deletedDateColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.DELETED_DATE);
+
+                    int databaseIdIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns._ID);
+
+                    Instance instance = new Instance.Builder()
+                            .displayName(cursor.getString(displayNameColumnIndex))
+                            .submissionUri(cursor.getString(submissionUriColumnIndex))
+                            .canEditWhenComplete(cursor.getString(canEditWhenCompleteIndex))
+                            .instanceFilePath(cursor.getString(instanceFilePathIndex))
+                            .jrFormId(cursor.getString(jrFormIdColumnIndex))
+                            .jrVersion(cursor.getString(jrVersionColumnIndex))
+                            .status(cursor.getString(statusColumnIndex))
+                            .lastStatusChangeDate(cursor.getLong(lastStatusChangeDateColumnIndex))
+                            .deletedDate(cursor.getLong(deletedDateColumnIndex))
+                            .databaseId(cursor.getLong(databaseIdIndex))
+                            .build();
+
+                    instances.add(instance);
+                }
+            } finally {
+                cursor.close();
+            }
+        }
+        return instances;
+    }
+
+    /**
+     * Returns the values of an instance as a ContentValues object for use with
+     * {@link org.odk.collect.android.dao.InstancesDao#saveInstance(ContentValues)} (ContentValues)}
+     * or {@link org.odk.collect.android.dao.InstancesDao#updateInstance(ContentValues, String, String[])}
+     *
+     * Does NOT include the database ID.
+     */
+    public static ContentValues getValuesFromInstanceObject(Instance instance) {
+        ContentValues values = new ContentValues();
+        values.put(InstanceProviderAPI.InstanceColumns.DISPLAY_NAME, instance.getDisplayName());
+        values.put(InstanceProviderAPI.InstanceColumns.SUBMISSION_URI, instance.getSubmissionUri());
+        values.put(InstanceProviderAPI.InstanceColumns.CAN_EDIT_WHEN_COMPLETE, instance.getCanEditWhenComplete());
+        values.put(InstanceProviderAPI.InstanceColumns.INSTANCE_FILE_PATH, instance.getInstanceFilePath());
+        values.put(InstanceProviderAPI.InstanceColumns.JR_FORM_ID, instance.getJrFormId());
+        values.put(InstanceProviderAPI.InstanceColumns.JR_VERSION, instance.getJrVersion());
+        values.put(InstanceProviderAPI.InstanceColumns.STATUS, instance.getStatus());
+        values.put(InstanceProviderAPI.InstanceColumns.LAST_STATUS_CHANGE_DATE, instance.getLastStatusChangeDate());
+        values.put(InstanceProviderAPI.InstanceColumns.DELETED_DATE, instance.getDeletedDate());
+
+        return values;
     }
 }
