+++ collect_app/src/main/java/org/odk/collect/android/tasks/FormLoaderTask.java
@@ -46,7 +46,6 @@ import org.odk.collect.android.logic.FileReferenceFactory;
 import org.odk.collect.android.logic.FormController;
 import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.android.utilities.FormDefCache;
-import org.odk.collect.android.utilities.DatabaseUtils;
 import org.odk.collect.android.utilities.ZipUtils;
 
 import java.io.File;
@@ -106,7 +105,7 @@ public class FormLoaderTask extends AsyncTask<String, String, FormLoaderTask.FEC
     FECWrapper data;
 
     public FormLoaderTask(String instancePath, String xpath, String waitingXPath) {
-        this.instancePath = DatabaseUtils.getAbsoluteFilePath(instancePath);
+        this.instancePath = FileUtils.getAbsoluteFilePath(instancePath);
         this.xpath = xpath;
         this.waitingXPath = waitingXPath;
     }
