--- collect_app/src/main/java/org/odk/collect/android/utilities/DatabaseUtils.java
@@ -1,169 +1,0 @@
-package org.odk.collect.android.utilities;
-
-import android.content.ContentValues;
-import android.database.Cursor;
-import android.provider.BaseColumns;
-
-import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.dto.Form;
-import org.odk.collect.android.dto.Instance;
-import org.odk.collect.android.provider.FormsProviderAPI;
-import org.odk.collect.android.provider.InstanceProviderAPI;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class DatabaseUtils {
-
-    private DatabaseUtils() { }
-
-    public static String getAbsoluteFilePath(String filePath) {
-        if (filePath == null) {
-            return "";
-        }
-        return filePath.startsWith(Collect.ODK_ROOT) ? filePath : Collect.ODK_ROOT + filePath;
-    }
-
-    public static String getRelativeFilePath(String filePath) {
-        if (filePath == null) {
-            return "";
-        }
-        return filePath.startsWith(Collect.ODK_ROOT) ? filePath.substring(Collect.ODK_ROOT.length()) : filePath;
-    }
-
-    /**
-     * Returns all instances available through the cursor and closes the cursor.
-     */
-    public static List<Instance> getInstancesFromCursor(Cursor cursor) {
-        List<Instance> instances = new ArrayList<>();
-        if (cursor != null) {
-            try {
-                cursor.moveToPosition(-1);
-                while (cursor.moveToNext()) {
-                    int displayNameColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.DISPLAY_NAME);
-                    int submissionUriColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.SUBMISSION_URI);
-                    int canEditWhenCompleteIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.CAN_EDIT_WHEN_COMPLETE);
-                    int instanceFilePathIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.INSTANCE_FILE_PATH);
-                    int jrFormIdColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.JR_FORM_ID);
-                    int jrVersionColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.JR_VERSION);
-                    int statusColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.STATUS);
-                    int lastStatusChangeDateColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.LAST_STATUS_CHANGE_DATE);
-                    int deletedDateColumnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.DELETED_DATE);
-
-                    int databaseIdIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns._ID);
-
-                    Instance instance = new Instance.Builder()
-                            .displayName(cursor.getString(displayNameColumnIndex))
-                            .submissionUri(cursor.getString(submissionUriColumnIndex))
-                            .canEditWhenComplete(cursor.getString(canEditWhenCompleteIndex))
-                            .instanceFilePath(cursor.getString(instanceFilePathIndex))
-                            .jrFormId(cursor.getString(jrFormIdColumnIndex))
-                            .jrVersion(cursor.getString(jrVersionColumnIndex))
-                            .status(cursor.getString(statusColumnIndex))
-                            .lastStatusChangeDate(cursor.getLong(lastStatusChangeDateColumnIndex))
-                            .deletedDate(cursor.getLong(deletedDateColumnIndex))
-                            .databaseId(cursor.getLong(databaseIdIndex))
-                            .build();
-
-                    instances.add(instance);
-                }
-            } finally {
-                cursor.close();
-            }
-        }
-        return instances;
-    }
-
-    /**
-     * Returns the values of an instance as a ContentValues object for use with
-     * {@link org.odk.collect.android.dao.InstancesDao#saveInstance(ContentValues)} (ContentValues)}
-     * or {@link org.odk.collect.android.dao.InstancesDao#updateInstance(ContentValues, String, String[])}
-     *
-     * Does NOT include the database ID.
-     */
-    public static ContentValues getValuesFromInstanceObject(Instance instance) {
-        ContentValues values = new ContentValues();
-        values.put(InstanceProviderAPI.InstanceColumns.DISPLAY_NAME, instance.getDisplayName());
-        values.put(InstanceProviderAPI.InstanceColumns.SUBMISSION_URI, instance.getSubmissionUri());
-        values.put(InstanceProviderAPI.InstanceColumns.CAN_EDIT_WHEN_COMPLETE, instance.getCanEditWhenComplete());
-        values.put(InstanceProviderAPI.InstanceColumns.INSTANCE_FILE_PATH, instance.getInstanceFilePath());
-        values.put(InstanceProviderAPI.InstanceColumns.JR_FORM_ID, instance.getJrFormId());
-        values.put(InstanceProviderAPI.InstanceColumns.JR_VERSION, instance.getJrVersion());
-        values.put(InstanceProviderAPI.InstanceColumns.STATUS, instance.getStatus());
-        values.put(InstanceProviderAPI.InstanceColumns.LAST_STATUS_CHANGE_DATE, instance.getLastStatusChangeDate());
-        values.put(InstanceProviderAPI.InstanceColumns.DELETED_DATE, instance.getDeletedDate());
-
-        return values;
-    }
-
-    /**
-     * Returns all forms available through the cursor and closes the cursor.
-     */
-    public static List<Form> getFormsFromCursor(Cursor cursor) {
-        List<Form> forms = new ArrayList<>();
-        if (cursor != null) {
-            try {
-                cursor.moveToPosition(-1);
-                while (cursor.moveToNext()) {
-                    int idColumnIndex = cursor.getColumnIndex(BaseColumns._ID);
-                    int displayNameColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.DISPLAY_NAME);
-                    int descriptionColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.DESCRIPTION);
-                    int jrFormIdColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.JR_FORM_ID);
-                    int jrVersionColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.JR_VERSION);
-                    int formFilePathColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.FORM_FILE_PATH);
-                    int submissionUriColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.SUBMISSION_URI);
-                    int base64RSAPublicKeyColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.BASE64_RSA_PUBLIC_KEY);
-                    int md5HashColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.MD5_HASH);
-                    int dateColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.DATE);
-                    int jrCacheFilePathColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.JRCACHE_FILE_PATH);
-                    int formMediaPathColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.FORM_MEDIA_PATH);
-                    int languageColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.LANGUAGE);
-                    int autoSendColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.AUTO_SEND);
-                    int autoDeleteColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.AUTO_DELETE);
-                    int lastDetectedFormVersionHashColumnIndex = cursor.getColumnIndex(FormsProviderAPI.FormsColumns.LAST_DETECTED_FORM_VERSION_HASH);
-
-                    Form form = new Form.Builder()
-                            .id(cursor.getInt(idColumnIndex))
-                            .displayName(cursor.getString(displayNameColumnIndex))
-                            .description(cursor.getString(descriptionColumnIndex))
-                            .jrFormId(cursor.getString(jrFormIdColumnIndex))
-                            .jrVersion(cursor.getString(jrVersionColumnIndex))
-                            .formFilePath(cursor.getString(formFilePathColumnIndex))
-                            .submissionUri(cursor.getString(submissionUriColumnIndex))
-                            .base64RSAPublicKey(cursor.getString(base64RSAPublicKeyColumnIndex))
-                            .md5Hash(cursor.getString(md5HashColumnIndex))
-                            .date(cursor.getLong(dateColumnIndex))
-                            .jrCacheFilePath(cursor.getString(jrCacheFilePathColumnIndex))
-                            .formMediaPath(cursor.getString(formMediaPathColumnIndex))
-                            .language(cursor.getString(languageColumnIndex))
-                            .autoSend(cursor.getString(autoSendColumnIndex))
-                            .autoDelete(cursor.getString(autoDeleteColumnIndex))
-                            .lastDetectedFormVersionHash(cursor.getString(lastDetectedFormVersionHashColumnIndex))
-                            .build();
-
-                    forms.add(form);
-                }
-            } finally {
-                cursor.close();
-            }
-        }
-        return forms;
-    }
-
-    public static ContentValues getValuesFromFormObject(Form form) {
-        ContentValues values = new ContentValues();
-        values.put(FormsProviderAPI.FormsColumns.DISPLAY_NAME, form.getDisplayName());
-        values.put(FormsProviderAPI.FormsColumns.DESCRIPTION, form.getDescription());
-        values.put(FormsProviderAPI.FormsColumns.JR_FORM_ID, form.getJrFormId());
-        values.put(FormsProviderAPI.FormsColumns.JR_VERSION, form.getJrVersion());
-        values.put(FormsProviderAPI.FormsColumns.FORM_FILE_PATH, form.getFormFilePath());
-        values.put(FormsProviderAPI.FormsColumns.SUBMISSION_URI, form.getSubmissionUri());
-        values.put(FormsProviderAPI.FormsColumns.BASE64_RSA_PUBLIC_KEY, form.getBASE64RSAPublicKey());
-        values.put(FormsProviderAPI.FormsColumns.MD5_HASH, form.getMD5Hash());
-        values.put(FormsProviderAPI.FormsColumns.DATE, form.getDate());
-        values.put(FormsProviderAPI.FormsColumns.JRCACHE_FILE_PATH, form.getJrCacheFilePath());
-        values.put(FormsProviderAPI.FormsColumns.FORM_MEDIA_PATH, form.getFormMediaPath());
-        values.put(FormsProviderAPI.FormsColumns.LANGUAGE, form.getLanguage());
-        return values;
-    }
-}
