+++ collect_app/src/main/java/org/odk/collect/android/utilities/FormDownloader.java
@@ -302,8 +302,8 @@ public class FormDownloader {
 
     private Uri saveNewForm(Map<String, String> formInfo, File formFile, String mediaPath) {
         final ContentValues v = new ContentValues();
-        v.put(FormsProviderAPI.FormsColumns.FORM_FILE_PATH,          DatabaseUtils.getRelativeFilePath(formFile.getAbsolutePath()));
-        v.put(FormsProviderAPI.FormsColumns.FORM_MEDIA_PATH,         DatabaseUtils.getRelativeFilePath(mediaPath));
+        v.put(FormsProviderAPI.FormsColumns.FORM_FILE_PATH,          FileUtils.getRelativeFilePath(formFile.getAbsolutePath()));
+        v.put(FormsProviderAPI.FormsColumns.FORM_MEDIA_PATH,         FileUtils.getRelativeFilePath(mediaPath));
         v.put(FormsProviderAPI.FormsColumns.DISPLAY_NAME,            formInfo.get(FileUtils.TITLE));
         v.put(FormsProviderAPI.FormsColumns.JR_VERSION,              formInfo.get(FileUtils.VERSION));
         v.put(FormsProviderAPI.FormsColumns.JR_FORM_ID,              formInfo.get(FileUtils.FORMID));
@@ -356,7 +356,7 @@ public class FormDownloader {
                 FileUtils.deleteAndReport(f);
 
                 // set the file returned to the file we already had
-                String existingPath = DatabaseUtils.getAbsoluteFilePath(c.getString(c.getColumnIndex(FormsProviderAPI.FormsColumns.FORM_FILE_PATH)));
+                String existingPath = FileUtils.getAbsoluteFilePath(c.getString(c.getColumnIndex(FormsProviderAPI.FormsColumns.FORM_FILE_PATH)));
                 f = new File(existingPath);
                 Timber.w("Will use %s", existingPath);
             }
@@ -480,7 +480,7 @@ public class FormDownloader {
 
         private UriResult(Uri uri, String mediaPath, boolean isNew) {
             this.uri = uri;
-            this.mediaPath = DatabaseUtils.getAbsoluteFilePath(mediaPath);
+            this.mediaPath = FileUtils.getAbsoluteFilePath(mediaPath);
             this.isNew = isNew;
         }
